requirements.txt
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.7.0
Authlib==1.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
contourpy==1.3.1
cozepy==0.9.1
crcmod==1.7
cryptography==44.0.0
cycler==0.12.1
distlib==0.3.9
distro==1.9.0
filelock==3.16.1
fonttools==4.55.3
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jmespath==0.10.0
kiwisolver==1.4.8
matplotlib==3.10.0
alibabacloud-nls-python-sdk==1.0.2
numpy==1.24.3
openai==1.14.0
oss2==2.19.1
packaging==24.2
pillow==11.0.0
pipenv==2024.4.0
platformdirs==4.3.6
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
pygame==2.5.2
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
schedule==1.2.2
selenium>=4.0.0
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.4.6
soundfile==0.12.1
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.28.0
webdriver_manager>=4.0.0
